Replace Room and Retrofit with SqlDelight and Ktor (#3)
* Replace Room with SqlDelight, Retrofit with Ktor

* Re-add OkHttp interceptors

* Clean up

+++ api/src/main/java/com/hellmund/api/model/ApiGenre.kt
@@ -2,4 +2,4 @@ package com.hellmund.api.model
 
 data class GenresResponse(val genres: List<ApiGenre>)
 
-data class ApiGenre(val id: Int, val name: String)
+data class ApiGenre(val id: Long, val name: String)
